More secure access code handling
Access codes are now 8 digits, but treated more like passwords:
Trying an access code takes at least a second, which presents a
big hurdle to any attempt to brute force access codes. While the
access code is stored in plain text, it's random, so an attacker
with a copy of the DB wouldn't have anything of value anyway.

Session IDs are left as 22-digits (>128-bit entropy) and all
communication of these is moved to POST operations for more
security.